chore: refactor avail package
